requirements.txt
PyQt5==5.15.11
pillow==11.2.1
pytesseract==0.3.13
watchdog==6.0.0
requests==2.32.4
python-dotenv==1.1.1
pyinstaller
google-genai